changing type of variable n in src/daemon/hmmc2.c/main() to uint64_t to fix crashes when master sends more than 2GB of data to client

+++ src/daemon/hmmc2.c
@@ -165,7 +165,7 @@ int
 main(int argc, char *argv[])
 {
   int              i, j;
-  int              n;
+  uint64_t              n;
   int              eod;
   int              size;
   char            *seq;
